+++ src/http.ts
@@ -16,35 +16,88 @@
  */
 
 import * as express from 'express';
+import * as fs from 'fs-extra';
+import * as http from 'http';
+import * as https from 'https';
+import * as path from 'path';
+import { getEGOFolder } from './util';
 
+
+/**
+ * Options for 'createHttpServer()' function.
+ */
+export interface CreateHttpServerOptions {
+    /**
+     * Force HTTP or not.
+     */
+    forceHttp?: boolean;
+}
 
 /**
  * Result of 'createHttpServer()' function.
  */
 export interface CreateHttpServerResult {
     /**
-     * The host instance.
+     * The express instance.
      */
-    readonly host: express.Express;
+    readonly app: express.Express;
     /**
      * Is HTTPs or not.
      */
     readonly https: boolean;
+    /**
+     * The server instance.
+     */
+    readonly server: http.Server | https.Server;
 }
 
 
 /**
  * Creates a new HTTP server instance.
  *
+ * @param {CreateHttpServerOptions} [opts] Custom options.
+ *
  * @return {CreateHttpServerResult} The result.
  */
-export function createHttpServer(): CreateHttpServerResult {
-    let https = false;
+export function createHttpServer(opts?: CreateHttpServerOptions): CreateHttpServerResult {
+    if (!opts) {
+        opts = {} as any;
+    }
 
-    const HOST = express();
+    let isHTTPs: boolean;
+    let server: http.Server | https.Server | false;
+
+    const app = express();
+
+    if (!opts.forceHttp) {
+        const EGO_FOLDER = getEGOFolder();
+
+        const CERT_FILE = path.resolve(
+            path.join(EGO_FOLDER, 'server.cert')
+        );
+        const KEY_FILE = path.resolve(
+            path.join(EGO_FOLDER, 'server.key')
+        );
+
+        if (fs.existsSync(CERT_FILE) && fs.existsSync(KEY_FILE)) {
+            isHTTPs = true;
+
+            server = https.createServer({
+                cert: fs.readFileSync(CERT_FILE),
+                key: fs.readFileSync(KEY_FILE),
+            }, app);
+        }
+    }
+
+    if (false === server) {
+        isHTTPs = false;
+
+        server = http.createServer(app);
+    }
 
     return {
-        host: HOST,
-        https,
+        app,
+        https: isHTTPs,
+        server,
     };
 }